chapter2/ex22 All Tasks

+++ chapter2/ex22/student/main.cpp
@@ -6,7 +6,7 @@ int main() {
     const double POUNDS_PER_TON = 2205.0;
     cin >> bag_capacity;
 
-    double bags_needed = POUNDS_PER_TON / bag_capacity;
+    int bags_needed = POUNDS_PER_TON / bag_capacity;
 
     cout << bags_needed << endl;
     return 0;
